+++ packages/platform/src/platform.ts
@@ -143,6 +143,7 @@ export default plugin(platformId, {
     Unauthorized: '' as StatusCode, // 401
     TokenExpired: '' as StatusCode, // 401
     TokenNotActive: '' as StatusCode<{ notBefore: number }>, // 401
+    PasswordExpired: '' as StatusCode, // 401
     Conflict: '' as StatusCode, // 409
     ExpiredLink: '' as StatusCode,
     UnknownMethod: '' as StatusCode<{ method: string }>,